STY: moved code for DialogSelectScan in a separate file
